Uploading self to scoreboard viable

+++ ScoreBoard.cpp
@@ -34,7 +34,7 @@ void ScoreBoard::getScores()
 			i++;
 		}
 		if (i % 2 == 0)
-			mScores.insert(pair<string, double>(values[0], stod(values[1])));
+			mScores.insert(pair<string, int>(values[0], stoi(values[1])));
